+++ components/Buttons/src/MDCButton.m
@@ -20,6 +20,7 @@
 #import "MaterialShadowElevations.h"
 #import "MaterialShadowLayer.h"
 #import "MaterialShapeLibrary.h"
+#import "MaterialShapes.h"
 #import "MaterialTypography.h"
 #import "MaterialMath.h"
 #import <MDFTextAccessibility/MDFTextAccessibility.h>
@@ -116,13 +117,16 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
 @property(nonatomic, readonly, strong) MDCShapedShadowLayer *layer;
 @property(nonatomic, assign) BOOL accessibilityTraitsIncludesButton;
 @property(nonatomic, assign) BOOL enableTitleFontForState;
+@property(nonatomic, assign) UIEdgeInsets visibleAreaInsets;
 @property(nonatomic) UIEdgeInsets hitAreaInsets;
+@property(nonatomic, assign) UIEdgeInsets currentVisibleAreaInsets;
 @end
 
 @implementation MDCButton
 
 @synthesize mdc_overrideBaseElevation = _mdc_overrideBaseElevation;
 @synthesize mdc_elevationDidChangeBlock = _mdc_elevationDidChangeBlock;
+@synthesize visibleAreaInsets = _visibleAreaInsets;
 @dynamic layer;
 
 + (Class)layerClass {
@@ -303,6 +307,18 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
   [self updateShadowColor];
   [self updateBackgroundColor];
   [self updateBorderColor];
+
+  if (self.centerVisibleArea) {
+    UIEdgeInsets visibleAreaInsets = self.visibleAreaInsets;
+    if (!UIEdgeInsetsEqualToEdgeInsets(visibleAreaInsets, self.currentVisibleAreaInsets)) {
+      self.currentVisibleAreaInsets = visibleAreaInsets;
+      MDCRectangleShapeGenerator *shapeGenerator =
+          [self generateShapeWithCornerRadius:self.layer.cornerRadius
+                            visibleAreaInsets:visibleAreaInsets];
+      [self configureLayerWithShapeGenerator:shapeGenerator];
+    }
+  }
+
   if (!self.layer.shapeGenerator) {
     self.layer.shadowPath = [self boundingPath].CGPath;
   }
@@ -357,7 +373,7 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
 }
 
 - (CGSize)sizeThatFits:(CGSize)size {
-  CGSize givenSizeWithInsets = CGSizeShrinkWithInsets(size, self.visibleAreaInsets);
+  CGSize givenSizeWithInsets = CGSizeShrinkWithInsets(size, _visibleAreaInsets);
   CGSize superSize = [super sizeThatFits:givenSizeWithInsets];
 
   if (self.minimumSize.height > 0) {
@@ -373,7 +389,7 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
     superSize.width = MIN(self.maximumSize.width, superSize.width);
   }
 
-  CGSize adjustedSize = CGSizeExpandWithInsets(superSize, self.visibleAreaInsets);
+  CGSize adjustedSize = CGSizeExpandWithInsets(superSize, _visibleAreaInsets);
   return adjustedSize;
 }
 
@@ -983,8 +999,10 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
 }
 
 - (void)setShapeGenerator:(id<MDCShapeGenerating>)shapeGenerator {
-  if (!UIEdgeInsetsEqualToEdgeInsets(self.visibleAreaInsets, UIEdgeInsetsZero)) {
-    // When visibleAreaInsets is set, custom shapeGenerater is not allow to be set through setter.
+  if (!UIEdgeInsetsEqualToEdgeInsets(_visibleAreaInsets, UIEdgeInsetsZero) ||
+      self.centerVisibleArea) {
+    // When visibleAreaInsets or centerVisibleArea is set, custom shapeGenerater is not allow
+    // to be set through setter.
     return;
   }
 
@@ -1060,14 +1078,14 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
 
 #pragma mark - Visible area
 
-- (void)setVisibleAreaInsets:(UIEdgeInsets)visibleAreaInsets {
-  if (UIEdgeInsetsEqualToEdgeInsets(visibleAreaInsets, _visibleAreaInsets)) {
+- (void)setCenterVisibleArea:(BOOL)centerVisibleArea {
+  if (_centerVisibleArea == centerVisibleArea) {
     return;
   }
 
-  _visibleAreaInsets = visibleAreaInsets;
+  _centerVisibleArea = centerVisibleArea;
 
-  if (UIEdgeInsetsEqualToEdgeInsets(visibleAreaInsets, UIEdgeInsetsZero)) {
+  if (!centerVisibleArea && UIEdgeInsetsEqualToEdgeInsets(_visibleAreaInsets, UIEdgeInsetsZero)) {
     self.shapeGenerator = nil;
 
     if (_cornerRadiusObserverAdded) {
@@ -1077,8 +1095,10 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
       _cornerRadiusObserverAdded = NO;
     }
   } else {
-    [self configureLayerWithShapeGenerator:[self generateShapeWithCornerRadius:self.layer
-                                                                                   .cornerRadius]];
+    MDCRectangleShapeGenerator *shapeGenerator =
+        [self generateShapeWithCornerRadius:self.layer.cornerRadius
+                          visibleAreaInsets:self.visibleAreaInsets];
+    [self configureLayerWithShapeGenerator:shapeGenerator];
 
     if (!_cornerRadiusObserverAdded) {
       [self.layer addObserver:self
@@ -1090,6 +1110,60 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
   }
 }
 
+- (void)setVisibleAreaInsets:(UIEdgeInsets)visibleAreaInsets {
+  if (UIEdgeInsetsEqualToEdgeInsets(visibleAreaInsets, _visibleAreaInsets)) {
+    return;
+  }
+
+  _visibleAreaInsets = visibleAreaInsets;
+
+  if (UIEdgeInsetsEqualToEdgeInsets(visibleAreaInsets, UIEdgeInsetsZero) &&
+      !self.centerVisibleArea) {
+    self.shapeGenerator = nil;
+
+    if (_cornerRadiusObserverAdded) {
+      [self.layer removeObserver:self
+                      forKeyPath:NSStringFromSelector(@selector(cornerRadius))
+                         context:kKVOContextCornerRadius];
+      _cornerRadiusObserverAdded = NO;
+    }
+  } else {
+    MDCRectangleShapeGenerator *shapeGenerator =
+        [self generateShapeWithCornerRadius:self.layer.cornerRadius
+                          visibleAreaInsets:visibleAreaInsets];
+    [self configureLayerWithShapeGenerator:shapeGenerator];
+
+    if (!_cornerRadiusObserverAdded) {
+      [self.layer addObserver:self
+                   forKeyPath:NSStringFromSelector(@selector(cornerRadius))
+                      options:NSKeyValueObservingOptionNew
+                      context:kKVOContextCornerRadius];
+      _cornerRadiusObserverAdded = YES;
+    }
+  }
+}
+
+- (UIEdgeInsets)visibleAreaInsets {
+  if (!UIEdgeInsetsEqualToEdgeInsets(_visibleAreaInsets, UIEdgeInsetsZero)) {
+    // Use custom visibleAreaInsets value when user sets it.
+    return _visibleAreaInsets;
+  }
+
+  UIEdgeInsets visibleAreaInsets = UIEdgeInsetsZero;
+  if (self.centerVisibleArea) {
+    CGSize visibleAreaSize = [self sizeThatFits:CGSizeMake(CGFLOAT_MAX, CGFLOAT_MAX)];
+    CGFloat additionalRequiredHeight =
+        MAX(0, CGRectGetHeight(self.bounds) - visibleAreaSize.height);
+    CGFloat additionalRequiredWidth = MAX(0, CGRectGetWidth(self.bounds) - visibleAreaSize.width);
+    visibleAreaInsets.top = MDCCeil(additionalRequiredHeight * 0.5f);
+    visibleAreaInsets.bottom = additionalRequiredHeight - visibleAreaInsets.top;
+    visibleAreaInsets.left = MDCCeil(additionalRequiredWidth * 0.5f);
+    visibleAreaInsets.right = additionalRequiredWidth - visibleAreaInsets.left;
+  }
+
+  return visibleAreaInsets;
+}
+
 - (void)observeValueForKeyPath:(NSString *)keyPath
                       ofObject:(id)object
                         change:(NSDictionary *)change
@@ -1097,28 +1171,29 @@ static NSAttributedString *uppercaseAttributedString(NSAttributedString *string)
   if (context == kKVOContextCornerRadius) {
     if (!UIEdgeInsetsEqualToEdgeInsets(self.visibleAreaInsets, UIEdgeInsetsZero) &&
         self.shapeGenerator) {
-      [self
-          configureLayerWithShapeGenerator:[self generateShapeWithCornerRadius:self.layer
-                                                                                   .cornerRadius]];
+      MDCRectangleShapeGenerator *shapeGenerator =
+          [self generateShapeWithCornerRadius:self.layer.cornerRadius
+                            visibleAreaInsets:self.visibleAreaInsets];
+      [self configureLayerWithShapeGenerator:shapeGenerator];
     }
   } else {
     [super observeValueForKeyPath:keyPath ofObject:object change:change context:context];
   }
 }
 
-- (MDCRectangleShapeGenerator *)generateShapeWithCornerRadius:(CGFloat)cornerRadius {
+- (MDCRectangleShapeGenerator *)generateShapeWithCornerRadius:(CGFloat)cornerRadius
+                                            visibleAreaInsets:(UIEdgeInsets)visibleAreaInsets {
   MDCRectangleShapeGenerator *shapeGenerator = [[MDCRectangleShapeGenerator alloc] init];
   MDCCornerTreatment *cornerTreatment =
       [[MDCRoundedCornerTreatment alloc] initWithRadius:cornerRadius];
   [shapeGenerator setCorners:cornerTreatment];
-  shapeGenerator.topLeftCornerOffset =
-      CGPointMake(self.visibleAreaInsets.left, self.visibleAreaInsets.top);
+  shapeGenerator.topLeftCornerOffset = CGPointMake(visibleAreaInsets.left, visibleAreaInsets.top);
   shapeGenerator.topRightCornerOffset =
-      CGPointMake(-self.visibleAreaInsets.right, self.visibleAreaInsets.top);
+      CGPointMake(-visibleAreaInsets.right, visibleAreaInsets.top);
   shapeGenerator.bottomLeftCornerOffset =
-      CGPointMake(self.visibleAreaInsets.left, -self.visibleAreaInsets.bottom);
+      CGPointMake(visibleAreaInsets.left, -visibleAreaInsets.bottom);
   shapeGenerator.bottomRightCornerOffset =
-      CGPointMake(-self.visibleAreaInsets.right, -self.visibleAreaInsets.bottom);
+      CGPointMake(-visibleAreaInsets.right, -visibleAreaInsets.bottom);
   return shapeGenerator;
 }