+++ components/NavigationDrawer/src/private/MDCBottomDrawerContainerViewController.m
@@ -631,9 +631,11 @@ NSString *const kMDCBottomDrawerScrollViewAccessibilityIdentifier =
 - (void)viewWillLayoutSubviews {
   [super viewWillLayoutSubviews];
 
-  // Have _contentHeaderTopInset recalculated the next time self.contentHeaderTopInset is called so
-  // it has the correct value when the app is in iPad Slide Over.
-  _contentHeaderTopInset = NSNotFound;
+  if (self.adjustLayoutForIPadSlideOver) {
+    // Have _contentHeaderTopInset recalculated the next time self.contentHeaderTopInset is called
+    // so it has the correct value when the app is in iPad Slide Over.
+    _contentHeaderTopInset = NSNotFound;
+  }
 
   [self setupLayout];
   UIView *topView;